[CITE-174] Resolving PR Comments
