+++ citesphere/src/main/webapp/WEB-INF/views/auth/personalAccessToken.html
@@ -6,14 +6,14 @@
 
 <body>
 	<div layout:fragment="content">
-		<div>
-			<h3>New Personal Access Token created for user [[${userName}]] :
-				[[${tokenGenerated}]]</h3>
-		</div>
-
+		<p>
+			<a th:href="@{/auth/savePersonalAccessToken}"
+				class="btn btn-primary btn-sm" id="genPersonalToken"><i class="icon-circle-add"></i>
+				&nbsp;Generate New Token</a>
+		</p>
 		<div class="col-md-10">
 			<div>
-				<h4>Other tokens for this user:</h4>
+				<h4>[[${personalTokensCount}]] tokens for this user:</h4>
 			</div>
 			<table class="table table-striped table-bordered">
 				<tr>
