+++ citesphere/src/main/webapp/WEB-INF/views/layouts/main.html
@@ -104,8 +104,7 @@
 						<ul class="dropdown-menu">
 							<li role="presentation"><a th:href="@{/auth/tokens}">Authorized
 									OAuth Apps</a></li>
-							<li role="presentation"><a th:href="@{/auth/personalAccessToken}">Generate
-									Personal Access Token</a></li>
+							<li role="presentation"><a th:href="@{/auth/personalAccessToken}">Personal Access Tokens</a></li>
 						</ul></li>
 						
 					<li sec:authorize="isAuthenticated()" role="presentation">
